feat: refactor handlers
