--- src/error-handlers/500.js
@@ -1,12 +1,0 @@
-'use strict';
-
-module.exports = function (err, req, res) {
-  const error = err.message ? err.message : err;
-
-  const errorObj = {
-    status: 500,
-    message: error,
-  };
-
-  res.status(500).json(errorObj);
-};
