+++ src/handlers/not-found.js
@@ -1,6 +1,6 @@
 'use strict';
 
-function handle404(req, res) {
+function notFoundHandler(req, res) {
   const errorObj = {
     status: 404,
     message: 'Sorry, we could not find what you were looking for',
@@ -8,4 +8,4 @@ function handle404(req, res) {
   res.status(404).json(errorObj);
 }
 
-module.exports = handle404;
+module.exports = notFoundHandler;
